Add Domain Layer
Add the domain layer to map all the data results and operations through the domain.
Add domain models to be mapped by the data layer in the future
Add the repository interface to the domain to be implemented in the data layer and to be injected in the UseCase
Add the use case to get the pokedex list, to be injected through the viewmodel in the future
Create 2 helper clases. Failure.kt to handle possible failure scenarios and Either.kt to be able to calculate the possible result of the operator when the usecase is consumed. Right (Success) Left (Error)